refactor FeeCompatison code
